projects.js: 🔧 Fix

+++ src/server/projects.js
@@ -466,19 +466,26 @@ app.post(
 				profile.projects?.some((project) => String(project.id) === String(projectId))
 			);
 			if (!user)
-				return res.status(404).json({ ok: false, error: "Project not found" })
+				return res.status(404).json({ ok: false, error: "Project not found" });
 		}
 		if (!user?.projects?.find((p) => String(p.id) === String(projectId)) && !isDashTeam)
 			return res
 				.status(404)
 				.json({ ok: false, error: "Project not found in your profile" });
+
 		const userProjects = user?.projects;
 		const project = userProjects?.find(
 			(p) => String(p.id) === String(projectId)
 		) || null;
+		if (!project)
+			return res.status(404).json({ ok: false, error: "Project not found" });
 
 		if (project.thumbnailId > 1) {
-			await storage.deleteThumbnailFile(project.thumbnailId);
+			try {
+				await storage.deleteThumbnailFile(project.thumbnailId);
+			} catch (_) {
+				return res.status(500).json({ ok: false, error: "Failed to delete prev thumbnail" });
+			}
 		}
 
 		const thumbnailId = project.id;
